chore: extract Fleet code to a separate file
